# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/authentication/refresh-token": {
      "post": {
        "operationId": "BaseAuthenticationController_refreshToken",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "default": {
                    "value": {
                      "data": null
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "REFRESH Token": []
          }
        ]
      }
    },
    "/api/authentication/login/{roleId}": {
      "post": {
        "operationId": "BaseAuthenticationController_login",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailPasswordLoginDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/authentication/forget-password/{roleId}": {
      "post": {
        "operationId": "BaseAuthenticationController_forgetPassword",
        "parameters": [
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetPasswordDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/authentication/verify": {
      "post": {
        "operationId": "BaseAuthenticationController_verifyUser",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "VERIFY Token": []
          }
        ]
      }
    },
    "/api/authentication/verify-reset-password": {
      "post": {
        "operationId": "BaseAuthenticationController_verifyOtp",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "VERIFY Token": []
          }
        ]
      }
    },
    "/api/authentication/reset-password": {
      "post": {
        "operationId": "BaseAuthenticationController_resetPassword",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "PASSWORD_RESET Token": []
          }
        ]
      }
    },
    "/api/authentication/logout": {
      "post": {
        "operationId": "BaseAuthenticationController_logout",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/roles": {
      "get": {
        "operationId": "RoleController_get",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "all-roles": {
                    "summary": "All Roles",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": "string"
                        }
                      ]
                    }
                  },
                  "role-with-id": {
                    "summary": "Role with Id",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": "string",
                          "Permissions": [
                            {
                              "name": "John Doe",
                              "method": [
                                "string"
                              ]
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "post": {
        "operationId": "RoleController_post",
        "parameters": [
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
    "/api/roles/{id}": {
      "get": {
        "operationId": "RoleController_get",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "all-roles": {
                    "summary": "All Roles",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": "string"
                        }
                      ]
                    }
                  },
                  "role-with-id": {
                    "summary": "Role with Id",
                    "value": {
                      "data": [
                        {
                          "id": 1,
                          "name": "John Doe",
                          "default": "string",
                          "Permissions": [
                            {
                              "name": "John Doe",
                              "method": [
                                "string"
                              ]
                            }
                          ]
                        }
                      ]
                    }
                  }
                }
              }
            },
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      },
      "delete": {
        "operationId": "RoleController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "Locale",
            "schema": {
              "example": "en"
            }
          },
          {
            "in": "header",
            "required": false,
            "name": "isLocalized",
            "schema": {
              "example": false,
              "type": "boolean",
              "enum": [
                true,
                false
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ],
        "security": [
          {
            "ACCESS Token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Structure",
    "description": "https://xd.adobe.com/view/f4331c27-8826-49e3-8d40-3b6361c2a108-a160/",
    "version": "1.0",
    "contact": {
      "name": "Fahd Hakem",
      "url": "www.linkedin.com/in/fahd-hakem",
      "email": "fhakem75@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3030",
      "description": "Local"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Sandbox"
    },
    {
      "url": "http://127.0.0.1:3000",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "ACCESS Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "REFRESH Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "PASSWORD_RESET Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "VERIFY Token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EmailPasswordLoginDTO": {
        "type": "object",
        "properties": {
          "fcm": {
            "type": "string",
            "example": "user"
          },
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "password": {
            "type": "string",
            "example": "Default@123",
            "description": "Password must be: 8+ characters, 1 uppercase, 1 lowercase, 1 number, 1 special character"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ForgetPasswordDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@test.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "VerifyOtpDTO": {
        "type": "object",
        "properties": {
          "otp": {
            "type": "string",
            "example": "1234"
          }
        },
        "required": [
          "otp"
        ]
      },
      "ResetPasswordDTO": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "example": "Default@123",
            "description": "Password must be: 8+ characters, 1 uppercase, 1 lowercase, 1 number, 1 special character"
          }
        },
        "required": [
          "password"
        ]
      },
      "UpdateRoleDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": "{\"en\": \"John\", \"ar\": \"جون\"}"
          },
          "key": {
            "type": "string"
          },
          "permissionIds": {
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "permissionIds"
        ]
      },
      "CreateRoleDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "object",
            "example": "{\"en\": \"John\", \"ar\": \"جون\"}"
          },
          "key": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "key"
        ]
      },
...
    }
  }
}